Add: Start testing channel encode/decode

+++ src/bits.rs
@@ -2,11 +2,11 @@
 use bitvec::prelude::*;
 
 // Bit field where the MSB is stored first (0th)
-pub type Bits = BitVec<usize, Msb0>;
+pub type Bits = BitVec<u8, Msb0>;
 
 // Convenient macro that wraps bitvec! to create new Bits instances
 #[macro_export] macro_rules! new_bits {
     ($($elem:expr),* $(,)?) => {
-        bitvec::bitvec![usize, Msb0; $($elem),*]
+        bitvec::bitvec![u8, Msb0; $($elem),*]
     };
 }
